Make bspwm-panel.service static
Now it is required by bspwm; we don't have to specifically enable it
anymore (making things easier as we'll always use it with bspwm).

Note that because we only use "Requires" and not "After", bspwm will
continue launching whether or not the panel is ready (as the latter
requires xrdb to be loaded); thus we don't worry about xrdb throttling
our entire desktop.