Remove display-server from WantedBy= in sockets
Now xorg.socket and urxvtd.socket are only wanted by sockets.target;
this is what we should use when enabling sockets.  We add a Before=
to display-server.target so that xorg.socket will be ready for the
target if it is enabled (but will not be started by it if disabled).