Convert delay.service to xorg-delay@.service
Now we can fine-tune the amount of time we need to wait between Xorg
starting properly and launching other units.
We also start xorg eagerly, but we do not force this to happen before
x11.target in the unit alone; this is xorg-delay's responsibility by
starting after Xorg and before x11.target. This way, once we disable
xorg-delay@.service, we are left with Xorg not having to start before
x11.target (so other programs can launch earlier).