Have sxhkd.service launch rofi.service
sxhkd does not include a binding to use rofi, thus if we activate sxhkd
we also activate rofi as well. Note we still allow rofi to be enabled
just in case we explicitly need it in the future.
Doing this allows us to easily switch between window managers that do
not use rofi (for example, dwm).