Speed up EditorConfig with core C library
The python external (not the default) is quite fast, but the C library
is far better at parsing. It does, however, need an external library to
be installed, using the AUR or linuxbrew. Should it not be available, we
default back to the Python core.
Also removed ntpeters/vim-better-whitespace as EditorConfig can
automatically strip away whitespace.