[git] Automatically stash and rebase on git pull
On pull we automatically rebase, and on rebase we automatically stash
and then pop after the rebase is complete. Essentially we can pull on a
dirty working tree.
Also removed git watch and up; the latter can be replaced by installing
a gem called git-up, which can do the above should our git version be
too old.