[git] Updated prepare-commit-msg with better regex
The regular expression now used is much more specific than the previous
and as such will be less likely to trigger on unrelated branches.
It also encourages consistent name with branches (for example, no random
uppercase letters in the branch name).
The ticket format is also not exposed so that commit messages are again
consistent with Atlassian's format.