[git] Add hooks to automatically run ctags
On a commit, checkout, pull, rebase, or merge we run ctags on all
versioned files. This may result in multiple ctags processes running if
we use these operations too quickly and the codebase is large, but will
not interfere with the final output.

This was inspired/stolen from tpope's blog post 'Effortless Ctags with
Git'.