[tmux] Add plugins
The newest feature of which is the ability to copy things using
'<Prefix> + /' to search, and using vim keybindings to copy the
selection with 'y' to the system clipboard.
We can paste back into tmux with <Prefix> + ]', but I haven't yet found
a way to paste from the systemd clipboard if using urxvt; Ctrl+V or
something similar will work on more fully-featured terminal emulators.