[Git] Add commit message template
This includes line guides to help others create readable commit logs. It
may change in the future if I find it cumbersome to deal with.

It also incorporates a 'Resolves: ' field, which is where we now insert
a JIRA or GitHub ticket. In this fashion we do not waste needed space in
the header.