Add vim search plugins and adjust mappings
Grepper uses Neovim's asynchronous features to be faster than Unite
(which is now only used for VimFiler).
The Fzf binary is installed using a package manager; we download the
plugin only to keep our directory clean.
Mappings now attempt to be agnostic of the plugin driving it and more on
what it does to ease switching.