Spruce up Vim's bufferline with extra info
Now the buffer line is always shown even with only one buffer open as it
makes opening another feel like fewer things are rearranged.
It also shows the buffer number (so we can switch with `:b<num>`) along
with showing whether it has been modified so we know what to save.