Add aliases to enable/disable distcc in ccache
When we have only access to one machine (or the current one has a
different compiler) then distcc starts spewing out error messages. This
gives us the option to disable these options.

The only problem with this is that it assumes distcc is the only wrapper
for ccache, when others may potentially be used in tandem.