Do not recurse in generating tags file
Our default ctags behaviour is to recurse, but this is not necessary in
our ctags git hook as we are reading in the files from `git ls-files'.

We also remove the exclusion of the .git folder as it is not necessary
(since we already ignore it with the above command).