Prevent stow from folding symlinks
If we have a folder structure of stowable/a/b/c, and the folder ~/a does not
exist, then stow will create a symlink from stowable/a to ~/a. This is a
problem because anything placed in ~/a/ (or ~/a/b/, etc) will pop up in
the original stowable directory. Instead, the folder ~/a/b/ is created,
and a symlink is made to c.

This can also reduce our .gitignore file substantially.