Move git hooks into a global folder
This is a new feature in Git 2.9; we can have all our repositories use a
global git hooks folder. This is useful because we don't have to keep
calling `git init` every time we make a change to a hook.

Also move this folder out of '~/.config/git/templates'; that should only
contain actual template files (like the commit template).