Move git hooks into submodule
Note that we have not included test-build-before-commit, as the idea of
having a pre-commit hook that tests whether we broke the build may be a
bit flawed since it may take a long time to run (and commits should be
cheap; we do not want to discourage their use).

Hence, testing it probably won't be needed (but we'll see).