Update YouCompleteMe installation
We can call install.py with either Python 2.6+ or 3.3+.

Furthermore, we can no longer rely on Arch having the latest Boost
libraries; currently we are on 1.60 (which admittedly has been
recognised as outdated) while YouCompleteMe's version uses 1.6.1. Since
these libraries are pulled in anyway (unlike libclang which is
downloaded if we do not specify --system-libclang) it makes sense to use
them.