Prevent automatic opening of loclist in Neomake
Automatically opening it would cause window to shift upwards as room was
made for the loclist. This could be worked around with setting
'scrolloff=999', which would force the cursor and to stay in the same
place.

However, the loclist claims valuable room, and to keep things consistent
with YouCompleteMe (which doesn't open a loclist as it continuously
recompiles the file) we have disabled its opening. Since we will rarely
need to open it, using the default size of 10 (instead of 5) seems like
a low price to pay to avoid tweaking yet another configuration option.