Adjust Editorconfig's max line length requirement
Now if we edit a line that was already longer than the max line length,
we do not enter a new line to limit the line length. This will make
editing python files slightly less frustrating.

You can set formatoptions-=t to remove the hard requirement (enforcing
it when formatting with 'gq').

Also have Editorconfig highlight the column corresponding to the maximum
line length.